Ask for an object size in pixels, not a crop scale factor
The augmenter's crop was driven by size_factor, inverted into a crop
height. Nothing about that number said what would come out the other
side, so the only way to land objects at a usable size was to guess and
look at the patches.

min_long_axis/max_long_axis replace it. _crop_range measures the largest
labelled object with regionprops and computes the crop that resizes it
into the requested range, drawing log-uniformly so the large end is
actually visited -- RandomSizedCrop samples height uniformly and size is
the reciprocal, which piles every draw at the small end. It warns when
the range cannot be reached, and when objects would exceed half the
patch and be cut by the crop edge. Anchored on the largest object rather
than the median: extra small examples are harmless, missing large ones
are not.

Colour jitter and random gamma come with it, both off by default, since
single-channel images have no hue to shift. hue_limit bounds how far
colour can move, not how far it does.

ColorJitter converts RGB to HSV through OpenCV, which takes uint8 and
float32 and nothing else. normalize_intensity was handing back float64 --
np.percentile returns np.float64 scalars and dividing a float32 array by
one promotes the result -- so it now casts back, and augment() casts
whatever it is given.

+++ src/napari_ai_lab/utilities/dl_util.py
@@ -25,7 +25,12 @@ def normalize_intensity(image, intensity_low, intensity_high):
             intensity_high - intensity_low
         )
         image_float = np.clip(image_float, 0, 1)
-    return image_float
+    # Cast back rather than trusting the arithmetic: np.percentile hands back
+    # np.float64 scalars, and dividing a float32 array by one promotes the
+    # result to float64. That is twice the memory for no precision that
+    # matters, and it reaches OpenCV -- which accepts uint8 and float32 for
+    # RGB/HSV conversion and rejects float64 outright.
+    return image_float.astype(np.float32, copy=False)
 
 
 def compute_percentiles(image, percentile_low=1, percentile_high=99):
